Khan Re-Balancing (#292)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

<!-- NOTE: This format is NOT REQUIRED for things like runtime fixes,
code fixes and optimizations. In those instances you should try to give
a description of what is being fixed or optimized but are not required
to fill out the complete changelog. -->
<!-- Adjusting things like armor or weapon values for balance, while
they may be 'fixes' in their own way, are not considered code fixes as
described above and require the use of the Pull Request format below.-->


## About The Pull Request

Yeah, so - anyways.......

Khans had some issues with their balance. I've went through and re-did
some of their loadouts for their faction.
- The lever action has been removed from Enforcers, instead it is
replaced with the Auto-5. The Auto-5 is the alternative fit as it
doesn't fit in bags like the lever, holds one less round, but in return
fires slightly faster. This is on-par with actual other factions (as the
only faction that got lever shotguns round start are vet-legion. Auto-5s
are more common.)
- The Enforcer and Smith have kept the selectable trait book, but Khans
over-all had a silly amount of traits. Therefor Chemist and Senior
Enforcer had theirs removed. (Senior has enough perks, and the book for
Chemist made the loadout selection for perks worthless.)
- Smith has lost their explosive crafting, but Chemist has gotten the
advanced explosive crafting book akin to NCR sappers who get the choice
of one. (To make Chemist loadout balanced, the lowsurgery book has been
replaced with medium surgery, since both NCR medics and Legion aux get
mid-surgery, not low)
- Khan armors have had their wound protection updated. They are now
roughly equal to NCR/Legion armor counterparts.
- Khan senior enforcer has had his round-start caps lowered from 1,000
to 500. (Tbh this may still be a bit high but w/e I was suggested this
amount)
- Smith, in exchange for losing explosive trait book, has instead gained
more gun crafting recipes in exchange. Can make all base Khan guns plus
10mm smg and some others. Plus - can make alloys like NCR LO / Legion
forgemaster. (Lost their GnB part 4 server-start learned. Work for it,
like the LO has to.)
- Renamed Khan Smith into Khan Armorer, which is a canon name of a Khan
found in Fallout: NV. Minor change, but tbh fits the role better since
they are mostly focused on gun crafting _and_ smithing.
- Made the Senior Enforcer loadouts actually competitive with each
other. Hammer one now gets a deagle with it to provide a sidearm for
range, the 10mm had to be a meme since it was the WORST option possible
- and is now replaced with a Infiltrator since it's just a slightly
better R-91 with a scope. (I'd argue this makes up for the now lack of
an extra selectable trait for head enforcer.)

Also - small hotfix, why the fuck did the Senior Scribe had low surgery
skill but Scribe has mid-surgery??? Not khan related but come on.....
scribe gaming..........

## Why It's Good For The Game

The major factions on server have gotten a fair bit of balancing
attention but Khans lacked a lot, such as overlooked wound armor
protection, their round start caps, and how many traits they could get
(which honestly felt like an unintended oversight, maybe???)

I think this attempts to re-do the Khan loadouts without per-say
'nerfing' them but rather fixing the fact that many of the loadouts
aren't unique to the role (i.e easily replacing chemist with smith, or
vice-versa) - as this has some buffs as well in things Khans were weak
in, such as Smith not having much unique to them.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [ ] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->


## Changelog
<!-- This is mostly optional for small Pull Requests, but if you value
being credited for your work in-game be sure to fill it out. To opt-out,
remove everything below including the start and end :cl: brackets. -->

<!-- If your Pull Request includes a minor single line variable edit,
probably do not fill out this changelog. -->
<!-- However, if your pull request includes massive or immediately
noticeable changes, briefly describe those changes in some way in this
changelog. -->

:cl:
balance: See about section
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->