Bobstation Ghost Port (#300)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

<!-- NOTE: This format is NOT REQUIRED for things like runtime fixes,
code fixes and optimizations. In those instances you should try to give
a description of what is being fixed or optimized but are not required
to fill out the complete changelog. -->
<!-- Adjusting things like armor or weapon values for balance, while
they may be 'fixes' in their own way, are not considered code fixes as
described above and require the use of the Pull Request format below.-->


## About The Pull Request

Remembered Escape from Navarro and remembered Bobstation codebase had
interesting ghost mechanics with seeing in monochrome while dead and
de-limbed limbs pop off yourself and stay missing as a ghost.

Ended up finding out - most of it would port with ease! Sadly the motion
blur was really too annoying to get working, but I plan to look at it at
some point. And, while I could get monochrome working, I want to run
that by people first. I enjoy it, but I know it fucks with people. But
hey - lookin' gooood.

De-limbing makes the limb disappear on the ghost.

![image](https://github.com/f13babylon/f13babylon/assets/47883419/3b4c7ca3-5761-4159-aedd-e64ae4fc2aa7)

Wearing armor or clothing? Well, it'll stay on you when you're dead!

![image](https://github.com/f13babylon/f13babylon/assets/47883419/bfae701f-5931-4038-b52b-1a429e286882)

And even if you matrix it stays.

![image](https://github.com/f13babylon/f13babylon/assets/47883419/10e63ba4-e6a1-4e15-90da-a1ca82c50535)

## Why It's Good For The Game

It's more of a nice addition to how I even remember ollld TG worked with
ghosts, minus the delimbing and armor. The style kind of fits and makes
ghosts stand out, letting you see who they were - even if you didn't
catch the name - when alive.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [ ] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->


## Changelog
<!-- This is mostly optional for small Pull Requests, but if you value
being credited for your work in-game be sure to fill it out. To opt-out,
remove everything below including the start and end :cl: brackets. -->

<!-- If your Pull Request includes a minor single line variable edit,
probably do not fill out this changelog. -->
<!-- However, if your pull request includes massive or immediately
noticeable changes, briefly describe those changes in some way in this
changelog. -->

:cl:
add: Ports Bobstation's ghost system, preps for eventual porting of
motion blur code portion as well.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: maaacha <116771811+maaacha@users.noreply.github.com>
Co-authored-by: Tom <8881105+tf-4@users.noreply.github.com>