Fixes Representative Safe Cash (#322)
## About The Pull Request
A pile of cash in the NCR Rep safe had some messy code that made it
appear as caps even though it was stacked with NCR dollars. It should
presumably be NCR dollars, so now it is.

_(I have a feeling this was some more shitcode from me last year when I
helped build that base.)_
## Why It's Good For The Game
Bugfix good. Another Discord report. : )

## Pre-Merge Checklist
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.

## Changelog
:cl:
fix: Fixed NCR Rep safe contents.
/:cl: