Restores A Costume (#326)
## About The Pull Request

![easter](https://github.com/f13babylon/f13babylon/assets/132588088/026f6361-e8c6-4e94-bad0-f58563a355af)

The Easter Bunny costume was lazily removed from costume vendors servers
ago as it makes you really fast. What a shame for people who just
enjoyed the costume (me) – this PR restores it without the speed, but
adds a subtype that still has it if admins want to spawn it in and do
that.

## Why It's Good For The Game
More content is good! The costume existed anyway.

## Pre-Merge Checklist
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.


## Changelog

:cl:
add: Restored a deleted costume to costume vendors.
/:cl: