Fixes BoS Animal Pen (#328)
## About The Pull Request

![image](https://github.com/f13babylon/f13babylon/assets/132588088/288ef1c0-3938-4831-afa7-00a18a1e482b)

The side fences of the animal pens had no density and could be walked
through, sending animals rampant and upsetting the tin can men. Fixed
it.

## Why It's Good For The Game

A map change is soon but this is one of those really quick fixes in the
meantime. Another Discord report closed.

## Pre-Merge Checklist
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.

## Changelog

:cl:
fix: Fixed a mapping error with BoS animal pen.
/:cl: