Khan Loadout Fix (#334)
## About The Pull Request

The Khans haven't been able to get their unique items from the loadout,
because it would only give them to a job that no longer exists. This
fixes that.

## Why It's Good For The Game

Because we hate bugs! Well, this wasn't really a bug, but it was
overlooked and created problems.

## Pre-Merge Checklist
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.


## Changelog

:cl:
fix: Fixed Khans not being able to get their faction locked items from
the loadout.
/:cl: