Drip or Drown: Selectable Veteran Ranger Skins (#411)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

<!-- NOTE: This format is NOT REQUIRED for things like runtime fixes,
code fixes and optimizations. In those instances you should try to give
a description of what is being fixed or optimized but are not required
to fill out the complete changelog. -->
<!-- Adjusting things like armor or weapon values for balance, while
they may be 'fixes' in their own way, are not considered code fixes as
described above and require the use of the Pull Request format below.-->


## About The Pull Request

Veteran Rangers can select what type of helmet and suit they can wear
for the week. Be careful, as it cannot be undone once selected.


![dreamseeker_J5pQMpur0C](https://github.com/f13babylon/f13babylon/assets/93478960/8898d040-34b7-4ff2-9f8e-e6e9d8a2915f)

![dreamseeker_YTTEh5QQGs](https://github.com/f13babylon/f13babylon/assets/93478960/1bdf1e6a-c805-487e-9572-0bd40acba33a)


## Why It's Good For The Game

No more admin-only clothing. More drip options without adding in
unnecessary extra clothing items, only using icons. More varied Veteran
Rangers who have a specific look

I also kind of want to do this for some Legion armors too. 

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->


## Changelog
<!-- This is mostly optional for small Pull Requests, but if you value
being credited for your work in-game be sure to fill it out. To opt-out,
remove everything below including the start and end :cl: brackets. -->

<!-- If your Pull Request includes a minor single line variable edit,
probably do not fill out this changelog. -->
<!-- However, if your pull request includes massive or immediately
noticeable changes, briefly describe those changes in some way in this
changelog. -->

:cl:
add: Selectable Veteran Ranger skins
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->