Tuning up the README (#414)
Added some helpful 'for the badges' and updated our Is It Maintained?
segment! Also axed a few bits & added a disclaimer for Bethesda to cover
our hineys.

<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

<!-- NOTE: This format is NOT REQUIRED for things like runtime fixes,
code fixes and optimizations. In those instances you should try to give
a description of what is being fixed or optimized but are not required
to fill out the complete changelog. -->
<!-- Adjusting things like armor or weapon values for balance, while
they may be 'fixes' in their own way, are not considered code fixes as
described above and require the use of the Pull Request format below.-->


## About The Pull Request

Added some new badges, added a disclaimer, added NB & Fortuna to the
past servers - some other edits and some removals of redundant
information/distinguishing the 'us' from /tg/.

## Why It's Good For The Game

Helpful and organized!



## Changelog

Just some rewrites!


<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->